requirements.txt
aiohttp==3.9.3
attrs==23.2.0
botbuilder-core==4.15.0
botbuilder-integration-aiohttp==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0
botframework-streaming==4.15.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
idna==3.7
jsonpickle==1.4.2
multidict==6.0.5
openai==1.23.6
office365-rest-python-client
openpyxl
pandas
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
tenacity==8.5.0
typing_extensions==4.11.0
urllib3==1.26.18
Werkzeug==3.0.2
yarl==1.9.4
spacy==3.5.3
nltk==3.8.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
